Product Create with varation and variation options done

+++ database/migrations/2024_10_02_085602_create_product_variations_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(Product::class);
             $table->string('name');
-            $table->decimal('price', 8, 2)->nullable();
+            $table->integer('price')->nullable();
             $table->integer('quantity')->default(0);
             $table->timestamps();
         });
